requirements.txt
django
djangorestframework
Faker